Move UserOp.ts from test/ to clients/
